Update the profile menu to use the same component and allowed clicking on the profile itself. Closes #1099
