+++ src/ui/components/profile/stories/profile-item.stories.tsx
@@ -1,11 +1,9 @@
-import { ThemeProvider } from '@mui/material/styles';
 import { type Meta, type StoryObj } from '@storybook/react-webpack5';
 import React from 'react';
-import * as vi from 'vitest';
 
 import { consoleLog } from '@/shared/utils/console-log';
 
-import { ProfileItem } from '../profile-item';
+import { ProfileItemBase } from '../profile-item-base';
 
 const mockUserInfo = {
   nickname: 'Test User',
@@ -16,9 +14,9 @@ const mockUserInfo = {
   id: '1',
 };
 
-const meta: Meta<typeof ProfileItem> = {
-  title: 'Components/profile/ProfileItem',
-  component: ProfileItem,
+const meta: Meta<typeof ProfileItemBase> = {
+  title: 'Components/profile/ProfileItemBase',
+  component: ProfileItemBase,
   tags: ['autodocs'],
   parameters: {
     backgrounds: {
@@ -35,20 +33,20 @@ const meta: Meta<typeof ProfileItem> = {
   argTypes: {
     profileId: { control: 'text' },
     selectedProfileId: { control: 'text' },
-    switchAccount: { action: 'switched' },
+    onClick: { action: 'switched' },
     setLoadingId: { action: 'loadingSet' },
   },
 };
 
 export default meta;
 
-type Story = StoryObj<typeof ProfileItem>;
+type Story = StoryObj<typeof ProfileItemBase>;
 
 export const Selected: Story = {
   args: {
     profileId: '1',
     selectedProfileId: '1',
-    switchAccount: async (id) => consoleLog('Switch to', id),
+    onClick: async (id) => consoleLog('Switch to', id),
     setLoadingId: (id) => consoleLog('Set loading', id),
     userInfo: mockUserInfo,
   },
@@ -58,8 +56,19 @@ export const Unselected: Story = {
   args: {
     profileId: '2',
     selectedProfileId: '1',
-    switchAccount: async (id) => consoleLog('Switch to', id),
+    onClick: async (id) => consoleLog('Switch to', id),
     setLoadingId: (id) => consoleLog('Set loading', id),
     userInfo: mockUserInfo,
+  },
+};
+
+export const ActiveProfile: Story = {
+  args: {
+    profileId: '1',
+    selectedProfileId: '1',
+    onClick: async (id) => consoleLog('Switch to', id),
+    setLoadingId: (id) => consoleLog('Set loading', id),
+    userInfo: mockUserInfo,
+    activeProfileVariant: true,
   },
 };
