+++ src/ui/views/Dashboard/Components/Popup.tsx
@@ -25,12 +25,10 @@ import {
 import userCircleCheck from '@/ui/assets/svg/user-circle-check.svg';
 import userCirclePlus from '@/ui/assets/svg/user-circle-plus.svg';
 import { ProfileButton } from '@/ui/components/profile/profile-button';
-import { useUserInfo } from '@/ui/hooks/use-account-hooks';
+import { ProfileItem } from '@/ui/components/profile/profile-item';
 import { useProfiles } from '@/ui/hooks/useProfileHook';
 import vmsvg from 'ui/assets/svg/viewmore.svg';
 import { useWallet } from 'ui/utils';
-
-import { ProfileItem } from '../../../components/profile/profile-item';
 
 interface TransferConfirmationProps {
   isConfirmationOpen: boolean;
@@ -40,33 +38,9 @@ interface TransferConfirmationProps {
   userInfo: UserInfoResponse;
   current: WalletAccount;
   switchAccount: (profileId: string) => Promise<void>;
-  profileIds: string[];
+  profileIds?: string[];
   switchLoading: boolean;
 }
-
-const ProfileItemList = ({
-  profileId,
-  selectedProfileId,
-  switchAccount,
-  setLoadingId,
-}: {
-  profileId: string;
-  selectedProfileId: string;
-  switchAccount: (profileId: string) => Promise<void>;
-  setLoadingId: (id: string) => void;
-}) => {
-  const userInfo = useUserInfo(profileId);
-  return (
-    <ProfileItem
-      key={profileId}
-      profileId={profileId}
-      selectedProfileId={selectedProfileId}
-      switchAccount={switchAccount}
-      setLoadingId={setLoadingId}
-      userInfo={userInfo}
-    />
-  );
-};
 
 const Popup = (props: TransferConfirmationProps) => {
   const usewallet = useWallet();
@@ -137,10 +111,10 @@ const Popup = (props: TransferConfirmationProps) => {
               }}
             >
               {props.profileIds.map((profileId: string) => (
-                <ProfileItemList
+                <ProfileItem
                   key={profileId}
                   profileId={profileId}
-                  selectedProfileId={props.userInfo.id}
+                  selectedProfileId={props.userInfo?.id}
                   switchAccount={props.switchAccount}
                   setLoadingId={setLoadingId}
                 />
